#58 Moved Route Factory and cleaned up code
