+++ src/app/core/factories/index.ts
@@ -1,2 +1,3 @@
 export * from './abstract-poi-factory.service';
 export * from './outdoor-poi-factory.service';
+export * from './route-factory';
