+++ src/app/core/models/index.ts
@@ -3,3 +3,6 @@ export * from './campus';
 export * from './coordinates';
 export * from './outdoor-poi';
 export * from './poi';
+export * from './route';
+export * from './route-step';
+export * from './transport-mode';
