+++ src/app/core/services/index.ts
@@ -1,3 +1,4 @@
 export * from './google-apis.service';
 export * from './location.service';
 export * from './map.service';
+export * from './routes.service';
